Move compute_digit_distances_to_centroids to handwritten_digit_utils.py
